refactor - remove swagger from endpoint business logic into SwaggerService, Swagger page
