+++ client/src/components/containers/Endpoint/Endpoint.tsx
@@ -6,20 +6,13 @@ import ColumnsWithActions from "../../presentational/Endpoint/EndpointData";
 import MockifyLoader from "components/commons/Loader/MockifyLoader";
 import { EyeOutlined } from "@ant-design/icons";
 import ResourceService from "services/Resource";
-import {Drawer, Tabs } from "antd";
-import MockifyCodeEditor from "components/commons/CodeEditor/CodeEditor";
-
+import { Tabs } from "antd";
 const Endpoint : React.FC = () => {
   const { 
     endpoint, 
     loading,
     handleTabChange,
     key, 
-    handleCloseSwaggerDrawer,
-    handleShowSwaggerDrawer,
-    swaggerDocsCache,
-    selectedResourceSwaggerDocs,
-    swaggerDrawerVisible,
     selectedResource
   
   } = EndpointService();
@@ -41,7 +34,6 @@ const Endpoint : React.FC = () => {
           tabBarExtraContent={
             <MockifyButton 
             classes={['table-action-third', 'textAndIcon', ]} 
-            onClick={() => {handleShowSwaggerDrawer(); swaggerDocsCache();}} 
             icon={<EyeOutlined />}
             text={`Swagger 2.0 Docs`}
             />
@@ -62,16 +54,6 @@ const Endpoint : React.FC = () => {
             };
           })}
         />
-        <Drawer
-            title={`${selectedResource.resourceName} Swagger Docs`}
-            placement="right"
-            onClose={handleCloseSwaggerDrawer}
-            open={swaggerDrawerVisible}
-            width={600}
-          >
-            <MockifyCodeEditor 
-            value={selectedResourceSwaggerDocs} height={"auto"} width={"600px"} onChange={() => console} />
-          </Drawer>
         </>
       )
     }
